Bug 759885 - Add support for running mochitest-plain on B2G emulators, r=jmaher, DONTBUILD because NPOTB
UltraBlame original commit: e96e0eaa6d85c777d330682f56d79a3109de9c97

+++ testing/mochitest/runtestsb2g.py
@@ -40,10 +40,10 @@ class B2GOptions(MochitestOptions):
                     help = "host:port to use when connecting to Marionette")
         defaults["marionette"] = None
 
-        self.add_option("--emulator", action="store_true",
-                    dest = "emulator",
-                    help = "True if using a b2g emulator")
-        defaults["emulator"] = False
+        self.add_option("--emulator", action="store",
+                    type="string", dest = "emulator",
+                    help = "Architecture of emulator to use: x86 or arm")
+        defaults["emulator"] = None
 
         self.add_option("--adbpath", action="store",
                     type = "string", dest = "adbPath",
@@ -186,20 +186,37 @@ class B2GMochitest(Mochitest):
         self.remoteProfile = options.remoteTestRoot + '/profile'
         self._automation.setRemoteProfile(self.remoteProfile)
         self.remoteLog = options.remoteLogFile
-        self.remoteProfilesIniPath = '/data/b2g/mozilla/profiles.ini'
+        self.userJS = '/data/local/user.js'
+        self.testDir = '/data/local/tests'
+        self.remoteMozillaPath = '/data/b2g/mozilla'
+        self.remoteProfilesIniPath = os.path.join(self.remoteMozillaPath, 'profiles.ini')
         self.originalProfilesIni = None
 
     def cleanup(self, manifest, options):
-        self._dm.getFile(self.remoteLog, self.localLog)
-        self._dm.removeFile(self.remoteLog)
-        self._dm.removeDir(self.remoteProfile)
-
+        
         if self.originalProfilesIni:
             try:
-                self.restoreProfilesIni()
+                if not options.emulator:
+                    self.restoreProfilesIni()
                 os.remove(self.originalProfilesIni)
             except:
                 pass
+
+        if not options.emulator:
+            self._dm.getFile(self.remoteLog, self.localLog)
+            self._dm.removeFile(self.remoteLog)
+            self._dm.removeDir(self.remoteProfile)
+
+            
+            self._dm.checkCmdAs(['shell', 'rm', '-f', self.userJS])
+            if self._dm.useDDCopy:
+                self._dm.checkCmdAs(['shell', 'dd', 'if=%s.orig' % self.userJS, 'of=%s' % self.userJS])
+            else:
+                self._dm.checkCmdAs(['shell', 'cp', '%s.orig' % self.userJS, self.userJS])
+
+            
+            
+            self._automation.rebootDevice()
 
         if options.pidFile != "":
             try:
@@ -207,10 +224,6 @@ class B2GMochitest(Mochitest):
                 os.remove(options.pidFile + ".xpcshell.pid")
             except:
                 print "Warning: cleaning up pidfile '%s' was unsuccessful from the test harness" % options.pidFile
-
-        
-        
-        self._automation.rebootDevice()
 
     def findPath(self, paths, filename = None):
         for path in paths:
@@ -283,7 +296,6 @@ class B2GMochitest(Mochitest):
     def buildProfile(self, options):
         if self.localProfile:
             options.profilePath = self.localProfile
-        print 'buildProfile', repr(options)
         manifest = Mochitest.buildProfile(self, options)
         self.localProfile = options.profilePath
 
@@ -318,6 +330,10 @@ class B2GMochitest(Mochitest):
             config.write(configfile)
 
         self._dm.pushFile(newProfilesIni, self.remoteProfilesIniPath)
+        try:
+            os.remove(newProfilesIni)
+        except:
+            pass
 
     def buildURLOptions(self, options, env):
         self.localLog = options.logFile
@@ -334,12 +350,22 @@ class B2GMochitest(Mochitest):
         
         
         f = open(os.path.join(options.profilePath, "user.js"), "a")
-        f.write('user_pref("browser.homescreenURL", "data:text/html,mochitest-plain should start soon");\n')
+        f.write('user_pref("browser.homescreenURL", "data:text/html,<h1>mochitest-plain should start soon</h1>");\n')
         f.close()
 
+        
         self._dm.removeDir(self.remoteProfile)
         if self._dm.pushDir(options.profilePath, self.remoteProfile) == None:
             raise devicemanager.FileError("Unable to copy profile to device.")
+
+        
+        
+        self._dm.checkCmdAs(['shell', 'rm', '-f', '%s.orig' % self.userJS])
+        if self._dm.useDDCopy:
+            self._dm.checkCmdAs(['shell', 'dd', 'if=%s' % self.userJS, 'of=%s.orig' % self.userJS])
+        else:
+            self._dm.checkCmdAs(['shell', 'cp', self.userJS, '%s.orig' % self.userJS])
+        self._dm.pushFile(os.path.join(options.profilePath, "user.js"), self.userJS)
 
         self.updateProfilesIni(self.remoteProfile)
 
@@ -355,7 +381,10 @@ def main():
     options, args = parser.parse_args()
 
     
-    kwargs = {'emulator': options.emulator}
+    kwargs = {}
+    if options.emulator:
+        kwargs['emulator'] = options.emulator
+        auto.setEmulator(True)
     if options.b2gPath:
         kwargs['homedir'] = options.b2gPath
     if options.marionette:
@@ -382,6 +411,10 @@ def main():
     auto.setProduct("b2g")
 
     mochitest = B2GMochitest(auto, dm, options)
+
+    
+    
+    dm.mkDirs(mochitest.testDir)
 
     options = parser.verifyOptions(options, mochitest)
     if (options == None):
